Improve build script

+++ build.php
@@ -1,53 +1,63 @@
 <?php
 
-if (file_exists(__DIR__ . '/vendor/erusev/parsedown/Parsedown.php')) {
-    require __DIR__ . '/vendor/erusev/parsedown/Parsedown.php';
-} else {
-    die("Please do 'composer install' first!\n");
-}
-
 $pharName = "e621BRS";
 $srcRoot = realpath("src");
 $buildRoot = realpath("build");
 
-echo "Cleaning up...\n";
+if (ini_get("phar.readonly") == 0) {
+    echo "Building...\n";
 
-if (file_exists($buildRoot . '/' . $pharName . '.phar')) {
-    unlink($buildRoot . '/' . $pharName . '.phar');
+    echo " Cleaning up...\n";
+    if (file_exists($buildRoot . '/' . $pharName . '.phar')) {
+        unlink($buildRoot . '/' . $pharName . '.phar');
+    }
+
+    $phar = new Phar($buildRoot . "/" . $pharName . ".phar", FilesystemIterator::CURRENT_AS_FILEINFO, $pharName . ".phar");
+
+    echo " Building phar...\n";
+    $phar->buildFromDirectory($srcRoot, '/.php$/');
+    $phar->setStub($phar->createDefaultStub("run.php"));
+
+    echo " Copying additional files...\n";
+
+    if (file_exists($srcRoot . "/config.cfg.example")) {
+        copy($srcRoot . "/config.cfg.example", $buildRoot . "/config.cfg.example");
+    }
+
+    if (file_exists($srcRoot . "/run.bat")) {
+        copy($srcRoot . "/run.bat", $buildRoot . "/run.bat");
+    }
+
+    if (file_exists($srcRoot . "/run.sh")) {
+        copy($srcRoot . "/run.sh", $buildRoot . "/run.sh");
+    }
+
+    if (file_exists(__DIR__ . '/vendor/erusev/parsedown/Parsedown.php')) {
+        echo " Converting markdown files into HTML...\n";
+        require __DIR__ . '/vendor/erusev/parsedown/Parsedown.php';
+
+        $Parsedown = new Parsedown();
+        file_put_contents($buildRoot . "/README.html", $Parsedown->text(file_get_contents(__DIR__ . '/README.md')));
+        file_put_contents($buildRoot . "/LICENSE.html", $Parsedown->text(file_get_contents(__DIR__ . '/LICENSE.md')));
+        file_put_contents($buildRoot . "/CONTRIBUTING.html", $Parsedown->text(file_get_contents(__DIR__ . '/CONTRIBUTING.md')));
+    } else {
+        echo("! Can't parse markdown files into HTML, dependencies not installed - do 'composer install'!\n");
+    }
+
+    echo "Done!\n\n";
+} else {
+    echo "! Can't build - 'phar.readonly' is 'On', check php.ini!\n\n";
 }
-
-if (file_exists(__DIR__ . '/build.zip')) {
-    unlink(__DIR__ . '/build.zip');
-}
-
-echo "Building phar...\n";
-$phar = new Phar($buildRoot . "/" . $pharName . ".phar", FilesystemIterator::CURRENT_AS_FILEINFO, $pharName . ".phar");
-
-$phar->buildFromDirectory($srcRoot, '/.php$/');
-$phar->setStub($phar->createDefaultStub("run.php"));
-
-echo "Copying additional files...\n";
-
-if (file_exists($srcRoot . "/config.cfg.example")) {
-    copy($srcRoot . "/config.cfg.example", $buildRoot . "/config.cfg.example");
-}
-
-if (file_exists($srcRoot . "/run.bat")) {
-    copy($srcRoot . "/run.bat", $buildRoot . "/run.bat");
-}
-
-if (file_exists($srcRoot . "/run.sh")) {
-    copy($srcRoot . "/run.sh", $buildRoot . "/run.sh");
-}
-
-$Parsedown = new Parsedown();
-file_put_contents($buildRoot . "/README.html", $Parsedown->text(file_get_contents(__DIR__ . '/README.md')));
-file_put_contents($buildRoot . "/LICENSE.html", $Parsedown->text(file_get_contents(__DIR__ . '/LICENSE.md')));
-file_put_contents($buildRoot . "/CONTRIBUTING.html", $Parsedown->text(file_get_contents(__DIR__ . '/CONTRIBUTING.md')));
 
 if (file_exists(__DIR__ . "/build_template.zip")) {
     echo "Packing...\n";
 
+    echo " Cleaning up...\n";
+    if (file_exists(__DIR__ . '/build.zip')) {
+        unlink(__DIR__ . '/build.zip');
+    }
+
+    echo " Copying template...\n";
     copy(__DIR__ . "/build_template.zip", __DIR__ . "/build.zip");
 
     $zip = new ZipArchive;
@@ -66,6 +76,10 @@ if (file_exists(__DIR__ . "/build_template.zip")) {
 
         $zip->close();
     }
+
+    echo "Done!\n\n";
+} else {
+    echo "! Can't pack build, no build_template.zip available!\n\n";
 }
 
-echo "Finished!\n";
+echo "Finished!\n\n";
